requirements.txt
Django==4.1.2
djangorestframework==3.14.0
python-dotenv==0.21.0
django-filter==22.1
django-cors-headers==3.13.0
APScheduler==3.10.4
djangorestframework_simplejwt==5.2.2
requests==2.31.0
Pillow==9.5.0
